Distance filter has been added

+++ src/Entity/Dtos/AdvertReadDto.cs
@@ -43,9 +43,9 @@ namespace Entity.Dtos
         /// <summary>
         ///   Location Data
         /// </summary>
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
         public string City { get; set; }
 
@@ -53,7 +53,7 @@ namespace Entity.Dtos
 
         public string County { get; set; }
 
-        public int Distance { get; set; }
+        public double Distance { get; set; }
 
         public DateTime CreatedAt { get; set; }
